Error Handling

+++ src/main/java/jomily/anv/employees_skills/rest/EmployeeRestController.java
@@ -16,7 +16,7 @@ import jomily.anv.employees_skills.entity.Employee;
 import jomily.anv.employees_skills.service.EmployeeService;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/employees")
 public class EmployeeRestController {
 
 	// autowire the EmployeeService
@@ -24,20 +24,24 @@ public class EmployeeRestController {
 	private EmployeeService employeeService;
 
 	// add mapping for GET/employees
-	@GetMapping("/employees")
+	@GetMapping
 	public List<Employee> getEmployees() {
 		return employeeService.getEmployees();
 	}
 
 	// add mapping for GET/employees/{employeeId}
-	@GetMapping("/employees/{employeeId}")
+	@GetMapping("/{employeeId}")
 	public Employee getEmployee(@PathVariable int employeeId) {
 		Employee theEmployee = employeeService.getEmployee(employeeId);
+		
+		if(theEmployee==null) {
+			throw new EmployeeNotFoundException("Customer id not found - "+employeeId);
+		}
 		return theEmployee;
 	}
 
 	// add mapping for POST /employees -add a new employee
-	@PostMapping("/employees")
+	@PostMapping
 	public Employee addEmployee(@RequestBody Employee theEmployee) {
 		// set id to 0 just in case an id is passed... forcing insert instead of update
 		theEmployee.setId(0);
@@ -47,17 +51,20 @@ public class EmployeeRestController {
 	}
 
 	// add mapping for PUT /employees -update employee
-	@PutMapping("/employees")
+	@PutMapping
 	public Employee updateEmployee(@RequestBody Employee theEmployee) {
 		employeeService.saveEmployee(theEmployee);
 		return theEmployee;
 	}
 
 	// add mapping for DELETE / employees/{employeeId}
-	@DeleteMapping("/employees/{employeeId}")
+	@DeleteMapping("/{employeeId}")
 	public String deleteEmployee(@PathVariable int employeeId) {
 		Employee theEmployee = employeeService.getEmployee(employeeId);
-
+		
+		if(theEmployee==null) {
+			throw new EmployeeNotFoundException("Customer id not - "+employeeId);
+		}
 		employeeService.deleteEmployee(employeeId);
 		return "Deleted Employee id- " + employeeId;
 	}
