+++ src/main/java/jomily/anv/employees_skills/rest/SkillRestController.java
@@ -12,33 +12,35 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import jomily.anv.employees_skills.entity.Employee;
 import jomily.anv.employees_skills.entity.Skill;
 import jomily.anv.employees_skills.service.SkillService;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/skills")
 public class SkillRestController {
 
-	// autowire the SkillService
 	@Autowired
 	private SkillService skillService;
 
 	// add mapping for GET /skills
-	@GetMapping("/skills")
+	@GetMapping
 	public List<Skill> getSkills() {
 		return skillService.getSkills();
 	}
 
 	// add mapping for GET /skills/{skillId}
-	@GetMapping("/skills/{skillId}")
+	@GetMapping("/{skillId}")
 	public Skill getSkill(@PathVariable int skillId) {
 		Skill theSkill = skillService.getSkill(skillId);
+
+		if (theSkill == null) {
+			throw new SkillNotFoundException("Skill id not found - " + skillId);
+		}
 		return theSkill;
 	}
 
 	// add mapping for POST /skills -add a new skill
-	@PostMapping("/skills")
+	@PostMapping
 	public Skill addSkill(@RequestBody Skill theSkill) {
 		// set id to 0 just in case an id is passed... forcing insert instead of update
 		theSkill.setId(0);
@@ -48,17 +50,20 @@ public class SkillRestController {
 	}
 
 	// add mapping for PUT /skills -update skill
-	@PutMapping("/skills")
+	@PutMapping
 	public Skill updateSkill(@RequestBody Skill theSkill) {
 		skillService.saveSkill(theSkill);
 		return theSkill;
 	}
 
 	// add mapping for DELETE / skills/{skillId}
-	@DeleteMapping("/skills/{skillId}")
+	@DeleteMapping("/{skillId}")
 	public String deleteSkill(@PathVariable int skillId) {
 		Skill theSkill = skillService.getSkill(skillId);
 
+		if (theSkill == null) {
+			throw new SkillNotFoundException("Skill id not found - " + skillId);
+		}
 		skillService.deleteSkill(skillId);
 		return "Deleted Employee id- " + skillId;
 	}
